bringing back old configuration
